pyproject.toml
[project]
name = "fruit-classifier"
version = "0.1.0"
description = "Fruit classification with Fruits-360"
authors = [
    {name = "Daniil Knyazev"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "torch (>=2.10.0,<3.0.0)",
    "torchvision (>=0.25.0,<0.26.0)",
    "pytorch-lightning (>=2.6.1,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "hydra-colorlog (>=1.2.0,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "dvc (>=3.66.1,<4.0.0)",
    "dvc-gdrive (>=3.0.1,<4.0.0)",
    "mlflow (>=3.9.0,<4.0.0)",
    "onnx (>=1.20.1,<2.0.0)",
    "onnxruntime (>=1.24.1,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "pandas (<3)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "gdown (>=5.2.1,<6.0.0)",
    "onnxscript (>=0.6.2,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (<24)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]
